# Update sample entropy