--- a/tests/test_entropy.py
+++ b/tests/test_entropy.py
@@ -16,6 +16,8 @@ TS_SAMPLE_ENTROPY = [1, 4, 5, 1, 7, 3, 1, 2, 5, 8, 9, 7, 3, 7, 9, 5, 4, 3, 9, 1,
                      3, 5, 6, 7, 8, 5, 2, 8, 6, 3, 8, 2, 7, 1, 7, 3, 5, 6, 2, 1, 3, 7, 3, 5, 3, 7, 6, 7, 7, 2, 3, 1, 7,
                      8]
 
+np.random.seed(1234567)
+RANDOM_TIME_SERIES = np.random.rand(1000)
 
 class TestEntropy(unittest.TestCase):
     def test_shannonEntropyString(self):
@@ -25,17 +27,14 @@ class TestEntropy(unittest.TestCase):
         self.assertEqual(round(ent.shannon_entropy(TIME_SERIES), 5), SHANNON_ENTROPY)
 
     def test_sampleEntropy(self):
-        ts = TS_SAMPLE_ENTROPY
+        ts = np.array(TS_SAMPLE_ENTROPY)
         std_ts = np.std(ts)
-        sample_entropy = ent.sample_entropy(ts, 4, 0.2 * std_ts)
-        np.testing.assert_array_equal(np.around(sample_entropy, 8),
-                                      [2.21187685, 2.10787948, 2.36712361, 1.79175947])
+        sample_entropy = ent.sample_entropy(ts, 3, 0.2 * std_ts)
+        self.assertEqual(np.around(sample_entropy, 8), 1.79175947)
 
-    def test_multiScaleEntropy(self):
-        ts = TS_SAMPLE_ENTROPY
-        std_ts = np.std(ts)
-        multi_scale_entropy = ent.multiscale_entropy(ts, 4, 0.2 * std_ts)
-        np.testing.assert_array_equal(np.round(multi_scale_entropy, 4), np.array([2.2119, 2.6221, 1.7473, 1.9902]))
+    def test_multiScaleEntropy(self):  
+        multi_scale_entropy = ent.multiscale_entropy(RANDOM_TIME_SERIES, 3, maxscale = 4 )
+        np.testing.assert_array_equal(np.round(multi_scale_entropy, 8), np.array([2.52572864, 2.33537492, 1.65292302, 1.86075234]))
 
     def test_permutationEntropy(self):
         self.assertEqual(np.round(ent.permutation_entropy(TS_SAMPLE_ENTROPY, 3, 5), 4), 1.7120)