@jakobdreyer 2018-09-07
... hm ... There is a conflict with PY2.7. Will fix and make a new PR....